#49 Added a package.json template and a few scripts
The scripts "start" and "dev" can be run by typing "npm run start" or "npm start" (or the same with "dev") and they both start the node server. However, the script "db:fresh" can only be run by entering "npm run db:fresh" (I assume because of the colons [:]) and it's only purpose is to execute a file named "refreshDatabase.js", which will be implemented in the following commits.